+++ R/mcmc_diagnostics.R
@@ -15,13 +15,7 @@
 #' data.table::setDTthreads(1) # For CRAN
 #' mcmc_diagnostics(gaussian_example_fit)
 #'
-mcmc_diagnostics <- function(x, n) {
-  UseMethod("mcmc_diagnostics", x)
-}
-
-#' @export
-#' @rdname mcmc_diagnostics
-mcmc_diagnostics.dynamitefit <- function(x, n = 3L) {
+mcmc_diagnostics <- function(x, n = 3L) {
   stopifnot_(
     !missing(x),
     "Argument {.arg x} is missing."
@@ -37,7 +31,9 @@ mcmc_diagnostics.dynamitefit <- function(x, n = 3L) {
     ),
     "Argument {.arg n} must be a single {.cls integer}."
   )
-  if (!is.null(x$stanfit)) {
+  if (is.null(x$stanfit)) {
+    cat("No Stan model fit is available.")
+  } else {
     algorithm <- x$stanfit@stan_args[[1L]]$algorithm
     stopifnot_(
       algorithm %in% c("NUTS", "hmc"),
@@ -65,8 +61,6 @@ mcmc_diagnostics.dynamitefit <- function(x, n = 3L) {
     ][init, ]
     out <- matrix(rhat$rhat, dimnames = list(rhat$variable, ""))
     print(out, digits = 3)
-  } else {
-    cat("No Stan model fit is available.")
   }
   invisible(x)
 }
@@ -74,9 +68,14 @@ mcmc_diagnostics.dynamitefit <- function(x, n = 3L) {
 #' HMC Diagnostics for a Dynamite Model
 #'
 #' Prints the divergences, saturated treedepths, and low E-BFMI warnings.
+#'
 #' @export
 #' @family diagnostics
 #' @param x \[`dynamitefit`]\cr The model fit object.
+#' @return Returns `x` (invisibly).
+#' data.table::setDTthreads(1) # For CRAN
+#' hmc_diagnostics(gaussian_example_fit)
+#'
 hmc_diagnostics <- function(x) {
   stopifnot_(
     !missing(x),
@@ -86,53 +85,56 @@ hmc_diagnostics <- function(x) {
     is.dynamitefit(x),
     "Argument {.arg x} must be a {.cls dynamitefit} object."
   )
-  algorithm <- x$stanfit@stan_args[[1L]]$algorithm
-  stopifnot_(
-    algorithm %in% c("NUTS", "hmc"),
-    "MCMC diagnostics are only meaningful for samples from MCMC.
-    Model was estimated using the ", algorithm, "algorithm."
-  )
-
-  n_draws <- ndraws(x)
-  n_divs <- rstan::get_num_divergent(x$stanfit)
-  n_trees <- rstan::get_num_max_treedepth(x$stanfit)
-  bfmis <- rstan::get_bfmi(x$stanfit)
-  all_ok <- n_divs == 0L && n_trees == 0L && all(bfmis > 0.2)
-  cat("NUTS sampler diagnostics:\n")
-  all_ok_str <- ifelse_(
-    all_ok,
-    "\nNo divergences, saturated max treedepths or low E-BFMIs.\n",
-    ""
-  )
-  cat(all_ok_str)
-  div_str <- ifelse_(
-    n_divs > 0L,
-    paste0(
-      "\n", n_divs, " out of ", n_draws, " iterations ended with a ",
-      "divergence. See Stan documentation for details.\n"
-    ),
-    ""
-  )
-  cat(div_str)
-  mt <- x$stanfit@stan_args[[1L]]$control$max_treedepth
-  mt <- ifelse_(is.null(mt), 10, mt)
-  trees_str <- ifelse_(
-    n_trees > 0L,
-    paste0(
-      "\n", n_trees, " out of ", n_draws, " saturated the maximum ",
-      "tree depth of ", mt, ". See Stan documentation for details.\n"
-    ),
-    ""
-  )
-  cat(trees_str)
-  bfmis_str <- ifelse_(
-    any(bfmis < 0.2),
-    paste0(
-      "\nChain(s) ", cs(which(bfmis < 0.2)), " had E-BFMI below 0.2, ",
-      "indicating possible issues. See Stan documentation for details.\n"
-    ),
-    ""
-  )
-  cat(bfmis_str)
-
+  if (is.null(x$stanfit)) {
+    cat("No Stan model fit is available.")
+  } else {
+    algorithm <- x$stanfit@stan_args[[1L]]$algorithm
+    stopifnot_(
+      algorithm %in% c("NUTS", "hmc"),
+      "MCMC diagnostics are only meaningful for samples from MCMC.
+      Model was estimated using the ", algorithm, "algorithm."
+    )
+    n_draws <- ndraws(x)
+    n_divs <- rstan::get_num_divergent(x$stanfit)
+    n_trees <- rstan::get_num_max_treedepth(x$stanfit)
+    bfmis <- rstan::get_bfmi(x$stanfit)
+    all_ok <- n_divs == 0L && n_trees == 0L && all(bfmis > 0.2)
+    cat("NUTS sampler diagnostics:\n")
+    all_ok_str <- ifelse_(
+      all_ok,
+      "\nNo divergences, saturated max treedepths or low E-BFMIs.\n",
+      ""
+    )
+    cat(all_ok_str)
+    div_str <- ifelse_(
+      n_divs > 0L,
+      paste0(
+        "\n", n_divs, " out of ", n_draws, " iterations ended with a ",
+        "divergence. See Stan documentation for details.\n"
+      ),
+      ""
+    )
+    cat(div_str)
+    mt <- x$stanfit@stan_args[[1L]]$control$max_treedepth
+    mt <- ifelse_(is.null(mt), 10, mt)
+    trees_str <- ifelse_(
+      n_trees > 0L,
+      paste0(
+        "\n", n_trees, " out of ", n_draws, " saturated the maximum ",
+        "tree depth of ", mt, ". See Stan documentation for details.\n"
+      ),
+      ""
+    )
+    cat(trees_str)
+    bfmis_str <- ifelse_(
+      any(bfmis < 0.2),
+      paste0(
+        "\nChain(s) ", cs(which(bfmis < 0.2)), " had E-BFMI below 0.2, ",
+        "indicating possible issues. See Stan documentation for details.\n"
+      ),
+      ""
+    )
+    cat(bfmis_str)
+  }
+  invisible(x)
 }