+++ R/predict.R
@@ -164,6 +164,14 @@ predict.dynamitefit <- function(object, newdata = NULL,
                                 global_fixed = FALSE, n_draws = NULL, thin = 1,
                                 expand = TRUE, df = TRUE, ...) {
   stopifnot_(
+    !missing(object),
+    "Argument {.arg object} is missing."
+  )
+  stopifnot_(
+    is.dynamitefit(object),
+    "Argument {.arg object} must be a {.cls dynamitefit} object."
+  )
+  stopifnot_(
     !is.null(object$stanfit),
     "No Stan model fit is available."
   )